push iter_rows(_mut) into BaseSlice(Mut)
